Set up a very basic trigger for impulses (may not be the best way to do it, but it works). However, impulses seem to trigger around 45% (as shown on the chart) rather than the given value of 50%. Likewise, the period ends around 90% as show on the chart rather than 100%.
Since shock zones (reliant on "currentPercent" variable rather than threads directly) occur at the correct percentages on the graphs, my best guess is that whatever is causing the problem is either with the threads or the way the threads are being used by the rest of the program (or the GUI specifically). Maybe this is the same reason that the periods don't end automatically?

git-svn-id: file:///svn/CONGv2.0/trunk@1467 637b740e-b1db-41bf-a134-2c17f54ac410